+++ lib/magellan/gcs/proxy.rb
@@ -10,7 +10,9 @@ require 'magellan/gcs/proxy/expand_variable'
 require 'magellan/gcs/proxy/gcp'
 require 'magellan/gcs/proxy/log'
 require 'magellan/gcs/proxy/message_wrapper'
-require 'magellan/gcs/proxy/pubsub_logger'
+require 'magellan/gcs/proxy/progress_notification'
+require 'magellan/gcs/proxy/progress_notifier_adapter'
+require 'magellan/gcs/proxy/pubsub_progress_notifier'
 
 module Magellan
   module Gcs
