+++ lib/magellan/gcs/proxy/progress_notifier_adapter.rb
