+++ lib/magellan/gcs/proxy/pubsub_progress_notifier.rb
